requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
attrs==22.2.0
autopage==0.5.1
autopep8==2.0.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cliff==4.2.0
cmd2==2.4.3
cron-descriptor==1.2.35
cryptography==38.0.4
debtcollector==2.5.0
decorator==5.1.1
Django==4.1.2
django-celery-beat==2.5.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dogpile.cache==1.1.8
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
install==1.3.5
iso8601==1.1.0
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
keystoneauth1==5.1.0
kombu==5.2.4
msgpack==1.0.4
netaddr==0.8.0
netifaces==0.11.0
openstacksdk==1.0.1
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.7.0
oslo.config==9.0.0
oslo.context==5.1.1
oslo.i18n==5.1.0
oslo.log==5.2.0
oslo.serialization==5.0.0
oslo.utils==6.1.0
packaging==22.0
pbr==5.11.0
pkgutil_resolve_name==1.3.10
prettytable==3.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyinotify==0.9.6
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.2
python-cinderclient==9.3.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-glanceclient==4.2.0
python-keystoneclient==5.0.1
python-neutronclient==8.2.0
python-novaclient==18.1.0
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requestsexceptions==1.4.0
rfc3986==2.0.0
semantic-version==2.3.1
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
stevedore==4.1.1
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.13
vine==5.0.0
warlock==2.0.1
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.15.0
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.10.0
django-admin-numeric-filter==0.1.9
drf-spectacular==0.26.2